agent/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
groq>=0.9.0
mem0ai>=0.1.0
chromadb>=0.5.0
redis>=5.0.0
playwright>=1.44.0
pydantic>=2.7.0
send2trash>=1.8.3
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0
